+++ lib/opal/rspec.rb
@@ -1,20 +1,43 @@
 require 'opal'
 require 'opal/rspec/version'
+require 'opal/minitest'
 
 # Just register our opal code path with opal build tools
 Opal.append_path File.expand_path('../../../opal', __FILE__)
-Opal.append_path File.expand_path('../../../vendor_lib', __FILE__)
+
+# Catch our git submodule included directories
+%w{rspec rspec-core rspec-expectations rspec-mocks rspec-support}.each do |gem|
+  Opal.append_path File.expand_path("../../../#{gem}/lib", __FILE__)
+end
 
 Opal::Processor.dynamic_require_severity = :warning
 
-Opal::Processor.stub_file "rspec/matchers/built_in/have"
-Opal::Processor.stub_file "diff/lcs"
-Opal::Processor.stub_file "diff/lcs/hunk"
-Opal::Processor.stub_file "fileutils"
-Opal::Processor.stub_file "test/unit/assertions"
-Opal::Processor.stub_file "coderay"
-Opal::Processor.stub_file "optparse"
-Opal::Processor.stub_file "shellwords"
-Opal::Processor.stub_file "socket"
-Opal::Processor.stub_file "uri"
-Opal::Processor.stub_file "drb/drb"
+stubs = [
+  'mutex_m', # Used with some threading operations but seems to run OK without this
+  'prettyprint',
+  'tempfile', # Doesn't exist in Opal
+  'diff/lcs',
+  'diff/lcs/block',
+  'diff/lcs/callbacks',
+  'diff/lcs/change',
+  'diff/lcs/hunk',
+  'diff/lcs/internals',
+  'test/unit/assertions',
+
+  # Opal doesn't have optparse, yet
+  'optparse',
+
+  'shellwords',
+  'socket',
+  'uri',
+  'drb/drb',
+
+  # Minitest used to be in stdlib, now is in opal-minitest GEM,
+  # but this file does not exist
+  # (referenced from minitest_assertions_adapter.rb in RSpec)
+  'minitest/unit',
+
+  'cgi/util',
+]
+
+stubs.each {|mod| Opal::Processor.stub_file mod }