+++ opal/opal/rspec.rb
@@ -1,9 +1,4 @@
-class MiniTest
-  class Unit; end
-end
-
-Test = MiniTest
-
+require 'opal/rspec/pre_require_fixes'
 require 'opal/rspec/requires'
 require 'opal/rspec/fixes'
 require 'opal/rspec/text_formatter'
@@ -13,7 +8,7 @@ require 'opal/rspec/async'
 
 RSpec.configure do |config|
   # For now, always use our custom formatter for results
-  config.formatter = Opal::RSpec::Runner.default_formatter
+  config.default_formatter = Opal::RSpec::Runner.default_formatter
 
   # Async helpers for specs
   config.include Opal::RSpec::AsyncHelpers