+++ opal/opal/rspec/async.rb
@@ -98,8 +98,9 @@ module Opal
       #     end
       #   end
       #
-      def async(&block)
-        @example.continue_async(block)
+      def async(&block)        
+        # we may have timed out, in which case @example will be gone
+        @example.continue_async(block) if @example
       end
 
       # Runs the given block after a given duration. You are still required to
@@ -201,7 +202,8 @@ module Opal
       #
       # @see AsyncHelpers::ClassMethods.async
       def self.register(*args)
-        examples << new(*args)
+        example_group = args[0]
+        example_group.examples << new(*args)
       end
 
       # All async examples in specs.
@@ -211,10 +213,9 @@ module Opal
         @examples ||= []
       end
 
-      def run(example_group_instance, reporter, &after_run_block)
+      def run(example_group_instance, reporter)
         @example_group_instance = example_group_instance
         @reporter               = reporter
-        @after_run_block        = after_run_block
         @finished               = false
 
         should_wait = true
@@ -225,15 +226,22 @@ module Opal
         start(reporter)
 
         begin
-          run_before_each
+          run_before_example
           @example_group_instance.instance_exec(self, &@example_block)
+          if pending?
+            Pending.mark_fixed! self
+
+            raise Pending::PendingExampleFixedError,
+                  'Expected example to fail since it is pending, but it passed.',
+                  [location]
+          end
         rescue Exception => e
           set_exception(e)
           should_wait = false
         end
 
         if should_wait
-          delay options[:timeout] || 10 do
+          delay self.metadata[:timeout] || 10 do
             next if finished?
 
             set_exception RuntimeError.new("timeout")
@@ -264,7 +272,7 @@ module Opal
         @finished = true
 
         begin
-          run_after_each
+          run_after_example
         rescue Exception => e
           set_exception(e)
         ensure
@@ -282,7 +290,6 @@ module Opal
 
         finish(@reporter)
         ::RSpec.current_example = nil
-        @after_run_block.call
       end
     end
   end