+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.5.0 (edge)
+
+*   Update to RSpec 3.1 (core is 3.1.7, expectations/support 3.1.2, mocks 3.1.3)
+
+*   Remove copy of source (and just rely on git submodule fetch)
+
 ## 0.4.3 (2015-06-14)
 
 *   Allow the gem to be run under Opal 0.7 and 0.8